flytt validering av adopsjon på barnets alder-vilkår ut i egen fil
